lambda/requirements.txt
# 開発用依存関係
# 生成日時: $(date)

# 本番用依存関係
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
boto3==1.35.0

# 開発用依存関係
pytest==8.4.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
python-dotenv==1.0.1
ruff==0.5.5
mypy==1.11.0
boto3-stubs[essential]
types-python-dateutil
pre-commit==3.7.1

# 依存関係の依存関係
cachetools==5.5.2
pyasn1-modules==0.4.2
rsa==4.9.1
requests-oauthlib==2.0.0
httplib2==0.22.0
google-api-core==2.25.1
uritemplate==4.2.0
botocore==1.35.99
jmespath==1.0.1
s3transfer==0.10.4
python-dateutil==2.9.0.post0
urllib3==2.5.0
googleapis-common-protos==1.70.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyparsing==3.2.3
oauthlib==3.3.1
requests==2.32.4
six==1.17.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
